init `layout-config` lib file for `metadata` and `fonts`
